+++ src/main/java/com/pixel/synchronre/sychronremodule/model/dto/repartition/response/CalculRepartitionResp.java
@@ -2,11 +2,13 @@ package com.pixel.synchronre.sychronremodule.model.dto.repartition.response;
 
 import lombok.*;
 
+import java.math.BigDecimal;
+
 @Getter @Setter @NoArgsConstructor @AllArgsConstructor @Builder
 public class CalculRepartitionResp
 {
-    private Float capital;
-    private Float taux;
-    private Float tauxBesoinFac;
-    private Float besoinFacRestant;
+    private BigDecimal capital;
+    private BigDecimal taux;
+    private BigDecimal tauxBesoinFac;
+    private BigDecimal besoinFacRestant;
 }
